+++ ios/adyen_checkout/Sources/adyen_checkout/components/card/advanced/CardAdvancedComponent.swift
@@ -71,7 +71,11 @@ class CardAdvancedComponent: BaseCardComponent {
 //    }
 
     private func buildActionComponent(adyenContext: AdyenContext) -> AdyenActionComponent {
-        let actionComponent = AdyenActionComponent(context: adyenContext)
+        var configuration = AdyenActionComponent.Configuration()
+        if let threeDS2Config = cardComponentConfiguration?.threeDS2ConfigurationDTO {
+            configuration.threeDS = threeDS2Config.mapToThreeDS2Configuration()
+        }
+        let actionComponent = AdyenActionComponent(context: adyenContext, configuration: configuration)
         actionComponent.delegate = actionComponentDelegate
         actionComponent.presentationDelegate = getViewController()
         return actionComponent