+++ lib/src/drop_in/model/drop_in_configuration.dart
@@ -5,6 +5,8 @@ import 'package:adyen_checkout/src/common/model/payment_method_configurations/ca
 import 'package:adyen_checkout/src/common/model/payment_method_configurations/cash_app_pay_configuration.dart';
 import 'package:adyen_checkout/src/common/model/payment_method_configurations/google_pay/google_pay_configuration.dart';
 import 'package:adyen_checkout/src/common/model/payment_method_configurations/stored_payment_method_configuration.dart';
+import 'package:adyen_checkout/src/common/model/payment_method_configurations/twint_configuration.dart';
+import 'package:adyen_checkout/src/common/model/payment_method_configurations/three_ds2_configuration.dart';
 
 final class DropInConfiguration extends BaseConfiguration {
   final CardConfiguration? cardConfiguration;
@@ -12,6 +14,8 @@ final class DropInConfiguration extends BaseConfiguration {
   final GooglePayConfiguration? googlePayConfiguration;
   final CashAppPayConfiguration? cashAppPayConfiguration;
   final StoredPaymentMethodConfiguration? storedPaymentMethodConfiguration;
+  final TwintConfiguration? twintConfiguration;
+  final ThreeDS2Configuration? threeDS2Configuration;
   final bool skipListWhenSinglePaymentMethod;
   final String? preselectedPaymentMethodTitle;
   final Map<String, String>? paymentMethodNames;
@@ -27,6 +31,8 @@ final class DropInConfiguration extends BaseConfiguration {
     this.googlePayConfiguration,
     this.cashAppPayConfiguration,
     this.storedPaymentMethodConfiguration,
+    this.twintConfiguration,
+    this.threeDS2Configuration,
     this.preselectedPaymentMethodTitle,
     bool? skipListWhenSinglePaymentMethod,
     AnalyticsOptions? analyticsOptions,
@@ -42,6 +48,8 @@ final class DropInConfiguration extends BaseConfiguration {
         'googlePayConfiguration: $googlePayConfiguration, '
         'cashAppPayConfiguration: $cashAppPayConfiguration, '
         'storedPaymentMethodConfiguration: $storedPaymentMethodConfiguration, '
+        'twintConfiguration: $twintConfiguration, '
+        'threeDS2Configuration: $threeDS2Configuration, '
         'skipListWhenSinglePaymentMethod: $skipListWhenSinglePaymentMethod, '
         'preselectedPaymentMethodTitle: $preselectedPaymentMethodTitle, '
         'paymentMethodNames: $paymentMethodNames)';